Do not inline fallback impl
The fallback impl tried to confuse the optimizer, however as shown by
Travis (https://travis-ci.org/cesarb/clear_on_drop/jobs/268508682) this
is not always successful.

Make it a never-inline function instead.

It's possible that the #[inline(never)] is enough to fool the optimizer,
but for now keep the dummy atomic variable.